\docType{package}
\name{cbmodels-package}
\alias{cbmodels-package}
\title{Collection of mainly combined models}
\description{
  This package provides a collection of models, including
  "combined" models:
}
\details{
  \itemize{ \item PCA-LDA \item PLS-LDA \item PCA with
  added reference \item EMSC }
}

