\name{center}
\alias{center}
\title{center}
\usage{
  center(object, ...)
}
\arguments{
  \item{object}{the model}

  \item{...}{further argument to the overloaded functions.}
}
\description{
  helper function returning the centering parameters of
  models.
}
\author{
  C. Beleites
}
\seealso{
  \code{\link{pcalda}}, \code{\link{plslda}}
}

